+++ examples/basic-app/src-tauri/src/lib.rs
@@ -26,16 +26,15 @@ fn native_crash() {
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
-    let client = sentry::init((
-        "https://233a45e5efe34c47a3536797ce15dafa@o447951.ingest.sentry.io/5650507",
-        sentry::ClientOptions {
-            release: sentry::release_name!(),
-            debug: true,
-            ..Default::default()
-        },
-    ));
+    let options = sentry::ClientOptions::new()
+        .dsn("https://233a45e5efe34c47a3536797ce15dafa@o447951.ingest.sentry.io/5650507")
+        .maybe_release(sentry::release_name!())
+        .debug(true);
 
-    let _guard = tauri_plugin_sentry::minidump::init(&client);
+    #[cfg(any(target_os = "linux", target_os = "macos", target_os = "windows"))]
+    let options = options.add_integration(sentry::integrations::minidump::MinidumpIntegration::new());
+
+    let client = sentry::init(options);
 
     tauri::Builder::default()
         .plugin(tauri_plugin_sentry::init(&client))