+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- feat: Replace `sentry-rust-minidump` with `sentry-minidump`. The `minidump` feature now enables the `minidump` feature of `sentry`. Add `sentry::integrations::minidump::MinidumpIntegration` to your `sentry::ClientOptions` instead of calling `tauri_plugin_sentry::minidump::init`. The `tauri_plugin_sentry::minidump` re-export is removed. Minidumps are only available on Linux, macOS and Windows.
+
 ## v0.5.0
 
 - feat: Update `sentry` to v0.42 and `sentry-rust-minidump` to v0.13